requirements.txt
asgiref==3.8.1
attrs==24.2.0
azure-ai-textanalytics==5.3.0
azure-common==1.1.28
azure-core==1.32.0
azure-search-documents==11.5.2
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
coverage==7.6.7
Django==5.0.9
django-cors-headers==4.6.0
django-filter==24.3
django-mssql-backend==2.8.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
isodate==0.7.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mssql-django==1.5
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
PyJWT==2.10.0
pyodbc==5.2.0
pytest==8.3.3
pytest-cov==6.0.0
pytest-django==4.9.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.21.0
six==1.16.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
